backend/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
boto3
python-dotenv
retina-face==0.0.16
ultralytics==8.2.28
scikit-learn==1.4.2
facenet-pytorch==2.5.3
scikit-image==0.23.2
face_recognition==1.3.0
opencv-python-headless==4.9.0.80
dill==0.3.8
cvzone==1.6.1
gunicorn
dlib
h5py